requirements.txt
accelerate==1.1.1
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
openai==1.55.3
packaging==24.2
psutil==6.1.0
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers.git@c24c79ebf91f6f04faf287997848ed6e64d78899
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
librosa>=0.10.0
flask-cors>=4.0.0
